SDBELGA-389 Join belga_remote_newsml12 and belganewsml12 feed parsers into one
