+++ server/belga/io/feed_parsers/__init__.py
@@ -10,4 +10,3 @@ from . import belga_ats_newsml_1_2  # noqa
 from . import belga_kyodo_newsml_1_2  # noqa
 from . import belga_spreadsheet  # noqa
 from . import belga_tip_newsml_1_2  # noqa
-from . import belga_remote_newsml_1_2  # noqa
