+++ server/belga/io/feed_parsers/belga_newsml_1_2.py
@@ -8,12 +8,28 @@
 # AUTHORS and LICENSE files distributed with this source code, or
 # at https://www.appendsourcefabric.org/superdesk/license
 
+import os
+import hashlib
 import logging
+from io import BytesIO
 from copy import deepcopy
+from uuid import uuid4
+from ftplib import error_perm
+from tempfile import gettempdir
+from datetime import datetime
+
+from flask import current_app as app
+from xml.etree import ElementTree
+
+from superdesk import get_resource_service
+from superdesk.ftp import ftp_connect
 from superdesk.errors import ParserError
 from superdesk.io.registry import register_feed_parser
+from superdesk.io.feeding_services import FileFeedingService, FTPFeedingService
 from superdesk.publish.formatters.newsml_g2_formatter import XML_LANG
+from superdesk.media.media_operations import process_file_from_stream
 from superdesk.utc import local_to_utc
+
 from .base_belga_newsml_1_2 import BaseBelgaNewsMLOneFeedParser, SkipItemException
 
 
@@ -27,7 +43,12 @@ class BelgaNewsMLOneFeedParser(BaseBelgaNewsMLOneFeedParser):
 
     label = 'Belga News ML 1.2 Parser'
 
-    SUPPORTED_ASSET_TYPES = ('ALERT', 'SHORT', 'TEXT', 'BRIEF')
+    SUPPORTED_ASSET_TYPES = ('ALERT', 'SHORT', 'TEXT', 'BRIEF', 'ORIGINAL')
+    ASSET_TYPES_WITH_ATTACHMENTS = ('IMAGE', 'SOUND', 'CLIP', 'COMPONENT')
+
+    def __init__(self):
+        super().__init__()
+        self._provider = None
 
     def can_parse(self, xml):
         """
@@ -69,27 +90,28 @@ class BelgaNewsMLOneFeedParser(BaseBelgaNewsMLOneFeedParser):
         :param provider:
         :return:
         """
+
+        self._provider = provider
+        if self._provider is None:
+            self._provider = {}
+
         try:
             self.root = xml
             self._items = []
             self._item_seed = {}
-
             # parser the NewsEnvelope element
             self._item_seed.update(
                 self.parser_newsenvelop(xml.find('NewsEnvelope'))
             )
-
             # parser the NewsItem element
             for newsitem_el in xml.findall('NewsItem'):
                 try:
                     self.parser_newsitem(newsitem_el)
                 except SkipItemException:
                     continue
-                # l_item.append(item)
             return self._items
-
         except Exception as ex:
-            raise ParserError.newsmlOneParserError(ex, provider)
+            raise ParserError.newsmlOneParserError(ex, self._provider)
 
     def parser_newsenvelop(self, envelop_el):
         """
@@ -154,13 +176,19 @@ class BelgaNewsMLOneFeedParser(BaseBelgaNewsMLOneFeedParser):
                         "scheme": "genre"
                     })
 
+            # parse attachment
+            self._item_seed.update(
+                self.parse_attachments(news_component_1)
+            )
+
             # NewsComponent 2nd level
             # NOTE: each NewsComponent of 2nd level is a separate item with unique GUID
             for news_component_2 in news_component_1.findall('NewsComponent'):
                 # guid
                 guid = news_component_2.attrib.get('Duid')
-                if guid is None:
-                    continue
+                # most probably it's belga remote
+                if guid is None or guid == '0':
+                    guid = str(uuid4())
 
                 # create an item
                 # deepcopy to avoid having a pointer to `subject`
@@ -169,6 +197,10 @@ class BelgaNewsMLOneFeedParser(BaseBelgaNewsMLOneFeedParser):
                 # NewsComponent
                 try:
                     self.parser_newscomponent(item, news_component_2)
+                    # SDBELGA-328
+                    if item.get('abstract'):
+                        abstract = '<p>' + item['abstract'] + '</p>'
+                        item['body_html'] = abstract + item['body_html']
                 except SkipItemException:
                     continue
 
@@ -518,6 +550,131 @@ class BelgaNewsMLOneFeedParser(BaseBelgaNewsMLOneFeedParser):
                 # city
                 if element.attrib.get('FormalName', '') == 'City' and element.attrib.get('Value'):
                     item.setdefault('extra', {})['city'] = element.attrib.get('Value')
+
+    def parse_attachments(self, news_component_1):
+        attachments = []
+        for news_component_2 in news_component_1.findall('NewsComponent'):
+            role_name = self._get_role(news_component_2)
+            if role_name and role_name.upper() not in self.SUPPORTED_ASSET_TYPES:
+                for newscomponent in news_component_2.findall('NewsComponent'):
+                    component_role = self._get_role(newscomponent)
+                    if component_role and component_role.upper() in self.ASSET_TYPES_WITH_ATTACHMENTS:
+                        attachment = self.parse_attachment(newscomponent)
+                        if attachment:
+                            attachments.append(attachment)
+                            # remove element to avoid parsing it as news item
+                            news_component_1.remove(news_component_2)
+        if attachments:
+            return {
+                'attachments': attachments,
+                'ednote': 'The story has {} attachment(s)'.format(len(attachments)),
+            }
+        return {}
+
+    def parse_attachment(self, newscomponent_el):
+        """
+        Parse attachment component, save it to storage and return attachment id
+
+        <NewsComponent Duid="0" xml:lang="nl">
+            <Role FormalName="Image"/>
+            <DescriptiveMetadata>
+                <Property FormalName="ComponentClass" Value="Image"/>
+            </DescriptiveMetadata>
+            <ContentItem Href="IMG_0182.jpg">
+                <Format FormalName="Jpeg"/>
+                <Characteristics>
+                    <SizeInBytes>2267043</SizeInBytes>
+                    <Property FormalName="Width" Value="4032"/>
+                    <Property FormalName="Height" Value="3024"/>
+                </Characteristics>
+            </ContentItem>
+        </NewsComponent>
+        """
+        content_item = newscomponent_el.find('ContentItem')
+        if content_item is None:
+            return
+
+        # avoid re-adding media after item is ingested
+        guid = hashlib.md5(ElementTree.tostring(content_item)).hexdigest()
+        attachment_service = get_resource_service('attachments')
+        old_attachment = attachment_service.find_one(req=None, guid=guid)
+        if old_attachment:
+            return {'attachment': old_attachment['_id']}
+
+        filename = content_item.attrib.get('Href')
+        if filename is None:
+            return
+
+        format_name = ''
+        format_el = content_item.find('Format')
+        if format_el is not None:
+            format_name = format_el.attrib.get('FormalName')
+
+        content = self._get_file(filename)
+        if not content:
+            return
+        _, content_type, metadata = process_file_from_stream(content, 'application/' + format_name)
+        content.seek(0)
+        media_id = app.media.put(content,
+                                 filename=filename,
+                                 content_type=content_type,
+                                 metadata=metadata,
+                                 resource='attachments')
+        try:
+            ids = attachment_service.post([{
+                'media': media_id,
+                'filename': filename,
+                'title': filename,
+                'description': 'belga remote attachment',
+                'guid': guid,
+            }])
+            return {'attachment': next(iter(ids), None)}
+        except Exception as ex:
+            app.media.delete(media_id)
+
+    def _get_role(self, newscomponent_el):
+        role = newscomponent_el.find('Role')
+        if role is not None:
+            return role.attrib.get('FormalName')
+
+    def _get_file(self, filename):
+        config = self._provider.get('config', {})
+        path = config.get('path', '')
+        file_dir = os.path.join(path, 'attachments')
+        file_path = os.path.join(file_dir, filename)
+        try:
+            if self._provider.get('feeding_service') == 'ftp':
+                file_path = self._download_file(filename, file_path, config)
+            with open(file_path, 'rb') as f:
+                content = f.read()
+                self._move_file(file_dir, filename, config)
+                return BytesIO(content)
+        except (FileNotFoundError, error_perm) as e:
+            logger.warning('File %s not found', file_path)
+        except Exception as e:
+            logger.error(e)
+
+    def _download_file(self, filename, file_path, config):
+        tmp_dir = os.path.join(gettempdir(), filename)
+        with ftp_connect(config) as ftp, open(tmp_dir, 'wb') as f:
+            ftp.retrbinary('RETR ' + file_path, f.write)
+            return tmp_dir
+
+    def _move_file(self, file_dir, filename, config):
+        if self._provider.get('feeding_service') == 'ftp':
+            with ftp_connect(config) as ftp:
+                if config.get('move', False):
+                    ftp_service = FTPFeedingService()
+                    move_path, _ = ftp_service._create_move_folders(config, ftp)
+                    ftp_service._move(
+                        ftp, os.path.join(file_dir, filename), os.path.join(move_path, filename),
+                        datetime.now(), False
+                    )
+        else:
+            file_service = FileFeedingService()
+            # move processed attachments to the same folder with XML
+            file_dir = os.path.dirname(file_dir)
+            file_service.move_file(file_dir, 'attachments/' + filename, self._provider)
 
 
 register_feed_parser(BelgaNewsMLOneFeedParser.NAME, BelgaNewsMLOneFeedParser())
